+++ .eslintrc.json
@@ -1,0 +1,19 @@
+{
+    "extends": "airbnb",
+    "parser": "babel-eslint",
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "react/forbid-prop-types": 0,
+        "react/destructuring-assignment": 0,
+        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
+        "import/no-named-as-default": 0,
+        "import/no-named-as-default-member": 0
+      },
+      "globals": {
+        "window": false,
+        "fetch": false,
+        "all": false
+      }
+}
