+++ .expo/packager-info.json
@@ -1,0 +1,5 @@
+{
+  "expoServerPort": 19000,
+  "packagerPort": 19001,
+  "packagerPid": 32836
+}
