+++ app.json
@@ -1,6 +1,5 @@
 {
   "expo": {
-    "sdkVersion": "21.0.0",
-    "privacy": "unlisted"
+    "sdkVersion": "27.0.0"
   }
 }
