+++ app/config/redux.js
@@ -1,19 +1,17 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
-import logger from 'redux-logger';
+// import logger from 'redux-logger';
 import createSagaMiddleware from 'redux-saga';
 
-import { QuarterPickerState, rootQuarterPickerSaga } from '../screens/QuarterPicker';
+import { QuarterPickerState, rootQuarterPickerSaga } from '../containers/QuarterPicker';
+import { TeamPickerState } from '../containers/TeamPicker';
+import { BoxScoreNavigator, BoxScoreState } from '../containers/BoxScore';
 
-import { TeamPickerState } from '../screens/TeamPicker';
-
-import { BoxScoreState } from '../screens/BoxScore';
-
-import Navigation from './routes';
+// import Navigation from './routes';
 
 const sagaMiddleware = createSagaMiddleware();
 
 const NavigationState = (state, action) => {
-  const newState = Navigation.router.getStateForAction(action, state);
+  const newState = BoxScoreNavigator.router.getStateForAction(action, state);
   return newState || state;
 };
 
@@ -28,7 +26,7 @@ const middleware = [sagaMiddleware];
 
 if (process.env.NODE_ENV === 'doesnt seem useful') {
   // 'development') {
-  middleware.push(logger);
+  //  middleware.push(logger);
 }
 
 const store = createStore(reducer, applyMiddleware(...middleware));
