+++ app/containers/BoxScore/index.js
@@ -1,0 +1,5 @@
+import { BoxScoreNavigator } from './routes';
+import { BoxScoreState } from './redux';
+
+export { BoxScoreState, BoxScoreNavigator };
+export default null;
