+++ app/containers/ScoreBoard/index.js
@@ -1,0 +1,3 @@
+import ScoreBoard from './ScoreBoard';
+
+export default ScoreBoard;
