+++ app/data/teams.js
@@ -1,11 +1,29 @@
 export default [
-  { number: 0, owner: 'Lev', name: 'Entwistlers', initials: 'LP' },
-  { number: 1, owner: 'Marc', name: 'Bombers', initials: 'MD' },
-  { number: 2, owner: 'Danny', name: 'Boston Massacre', initials: 'DL' },
-  { number: 3, owner: 'Chris L.', name: 'Camel Toes', initials: 'CL' },
-  { number: 4, owner: 'Michael', name: '', initials: 'MF' },
-  { number: 5, owner: 'Rocky', name: 'Smashers', initials: 'RF' },
-  { number: 6, owner: 'Chris F.', name: 'Philly Pimps', initials: 'CF' },
-  { number: 7, owner: 'Jesse', name: 'Fluffydogs', initials: 'JG' },
-  { number: 8, owner: 'Ananda', name: 'Anacondas', initials: 'AG' },
+  {
+    number: 0, owner: 'Lev', name: 'Entwistlers', initials: 'LP',
+  },
+  {
+    number: 1, owner: 'Marc', name: 'Bombers', initials: 'MD',
+  },
+  {
+    number: 2, owner: 'Danny', name: 'Boston Massacre', initials: 'DL',
+  },
+  {
+    number: 3, owner: 'Chris L.', name: 'Camel Toes', initials: 'CL',
+  },
+  {
+    number: 4, owner: 'Michael', name: '', initials: 'MF',
+  },
+  {
+    number: 5, owner: 'Rocky', name: 'Smashers', initials: 'RF',
+  },
+  {
+    number: 6, owner: 'Chris F.', name: 'Philly Pimps', initials: 'CF',
+  },
+  {
+    number: 7, owner: 'Jesse', name: 'Fluffydogs', initials: 'JG',
+  },
+  {
+    number: 8, owner: 'Ananda', name: 'Anacondas', initials: 'AG',
+  },
 ];
