+++ app/index.js
@@ -4,9 +4,10 @@ import EStyleSheet from 'react-native-extended-stylesheet';
 import { Provider } from 'react-redux';
 
 // import Sample from './screens/Sample';
-import { BoxScoreNavigator } from './config/routes';
+import { BoxScoreNavigator } from './containers/BoxScore';
 import store from './config/redux';
 import Header from './components/Header';
+// import ScoreBoard from './containers/ScoreBoard';
 
 const styles = EStyleSheet.create({
   container: {
@@ -22,6 +23,15 @@ EStyleSheet.build({
   //  outline: 1,
 });
 
+/*
+  <Provider store={store}>
+    <View style={styles.container}>
+      <StatusBar backgroundColor="#ff0000" barStyle="default" />
+      <Header />
+      <ScoreBoard />
+    </View>
+  </Provider>
+*/
 export default () => (
   <Provider store={store}>
     <View style={styles.container}>
