--- app/screens/BoxScore/index.js
@@ -1,5 +1,0 @@
-import BoxScore from './BoxScore';
-import { BoxScoreState } from './redux';
-
-export { BoxScoreState };
-export default BoxScore;
