--- app/screens/QuarterPicker/index.js
@@ -1,42 +1,0 @@
-import {
-  // public actions
-  quarterPickerInitialized,
-  quarterDataRequested,
-} from './redux/actions';
-
-import {
-  // public selectors
-  getSeason,
-  getQuarter,
-  getLineups,
-  getGameData,
-  getIsQuarterPickerLoading,
-  getIsQuarterPickerInitialized,
-  // public reducer
-  QuarterPickerState,
-} from './redux/reducers';
-
-import {
-  // public saga
-  rootQuarterPickerSaga,
-} from './redux/sagas';
-
-import QuarterPicker from './QuarterPicker';
-
-export {
-  // public actions
-  quarterPickerInitialized,
-  quarterDataRequested,
-  // public selectors
-  getSeason,
-  getQuarter,
-  getLineups,
-  getGameData,
-  getIsQuarterPickerLoading,
-  getIsQuarterPickerInitialized,
-  // public reducer
-  QuarterPickerState,
-  // public saga
-  rootQuarterPickerSaga,
-};
-export default QuarterPicker;
